ops-agent/pyproject.toml
[project]
name = "ops-agent"
version = "0.1.0"
description = "Ops AI Agent - Event-driven operations AI system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.0.0",
    "loguru>=0.7.0",
    "orjson>=3.10.0",
    "pyjwt>=2.10.0",
    "cryptography>=44.0.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "paramiko>=3.5.0",
    "kubernetes>=31.0.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
    "pgvector>=0.3.0",
    "poethepoet>=0.42.1",
    "python-multipart>=0.0.22",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "websockets>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "fakeredis[json]>=2.26.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "Pillow>=10.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring Docker services (PG + Redis)",
    "e2e: marks end-to-end tests requiring full stack",
    "slow: marks tests that are slow (VL model calls, etc.)",
]

[tool.ruff]
line-length = 100

[tool.poe.tasks]
dev = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8000"
test = "pytest tests/ -v"
"test:int" = "pytest -m integration -v"
lint = "ruff check src/ tests/"
format = "ruff format src/ tests/"
migrate = "alembic upgrade head"
"migrate:new" = "alembic revision --autogenerate"
